Refactored LessonPair
