+++ src/components/LessonCreationTool.js
@@ -1,13 +1,15 @@
 import React, { useState } from 'react'
 import LessonPair from './LessonPair';
 
-export default function LessonCreationTool() { 
+export default function LessonCreationTool() {
+
+    const [lessonTree, setlessonTree] = useState([])
     
     return (
         <div>
             <p>Lesson Creation</p>
-            <LessonPair videoType="Call" />
-            <LessonPair videoType="Analysis" />
+            <LessonPair changeLessontree={setlessonTree} />
+            {lessonTree}
         </div>
     )
 }
